expose get option in fileshareclient
