+++ pkg/azclient/fileshareclient/custom_test.go
@@ -71,6 +71,20 @@ func init() {
 				gomega.Expect(resourceList).To(gomega.BeNil())
 			})
 		})
+		ginkgo.When("get requests are raised", func() {
+			ginkgo.It("should not return error", func(ctx context.Context) {
+				newResource, err := realClient.Get(ctx, resourceGroupName, accountName, resourceName, nil)
+				gomega.Expect(err).NotTo(gomega.HaveOccurred())
+				gomega.Expect(newResource).NotTo(gomega.BeNil())
+			})
+		})
+		ginkgo.When("invalid get requests are raised", func() {
+			ginkgo.It("should return 404 error", func(ctx context.Context) {
+				newResource, err := realClient.Get(ctx, resourceGroupName, accountName, resourceName+"notfound", nil)
+				gomega.Expect(err).To(gomega.HaveOccurred())
+				gomega.Expect(newResource).To(gomega.BeNil())
+			})
+		})
 	}
 
 	beforeAllFunc = func(ctx context.Context) {
